Lua Extend

+++ src/EFCoreExtend/EFCache/BaseCache/Default/QueryCache.cs
@@ -8,7 +8,7 @@ namespace EFCoreExtend.EFCache.BaseCache.Default
 {
     public class QueryCache : QueryCacheBase
     {
-        protected readonly static string _keyExt = "__";
+        protected readonly static string _keyExt = "_";
         protected readonly ILocalCache _dictCache = new LocalCache();
 
         protected string GetCacheKeyExt(string cacheType)
@@ -21,22 +21,23 @@ namespace EFCoreExtend.EFCache.BaseCache.Default
             return GetCacheKeyExt(cacheType) + cacheKey;
         }
 
-        protected override bool TryGetCache<TRtn>(string cacheType, string cacheKey, out TRtn rtn)
+        protected override bool TryGetCache<TRtn>(string cacheType, string cacheKey, out TRtn rtn, Type rtnType = null)
         {
             return _dictCache.TryGetCache<TRtn>(GetCacheKey(cacheType, cacheKey), out rtn);
         }
 
-        protected override bool TryGetCacheAndUpdateCacheTime<TRtn>(string cacheType, string cacheKey, DateTime expiry, out TRtn rtn)
+        protected override bool TryGetCacheAndUpdateCacheTime<TRtn>(string cacheType, string cacheKey, DateTime expiry, 
+            out TRtn rtn, Type rtnType = null)
         {
             return _dictCache.TryGetCache<TRtn>(GetCacheKey(cacheType, cacheKey), out rtn, expiry);
         }
 
-        protected override void SetCache<TM>(string cacheType, string cacheKey, TM cacheModel)
+        protected override void SetCache<TM>(string cacheType, string cacheKey, TM cacheModel, Type rtnType = null)
         {
             _dictCache.SetCache(GetCacheKey(cacheType, cacheKey), cacheModel);
         }
 
-        protected override void SetCache<TM>(string cacheType, string cacheKey, TM cacheModel, DateTime expiry)
+        protected override void SetCache<TM>(string cacheType, string cacheKey, TM cacheModel, DateTime expiry, Type rtnType = null)
         {
             _dictCache.SetCache(GetCacheKey(cacheType, cacheKey), cacheModel, expiry);
         }
@@ -48,8 +49,11 @@ namespace EFCoreExtend.EFCache.BaseCache.Default
 
         public override void RemoveRange(string cacheType)
         {
-            string kt = GetCacheKeyExt(cacheType);
-            _dictCache.RemoveRange(_dictCache.Caches.Keys.Where(l => l.StartsWith(kt)).ToList());
+            if (_dictCache.Caches?.Count > 0)
+            {
+                string kt = GetCacheKeyExt(cacheType);
+                _dictCache.RemoveRange(_dictCache.Caches.Keys.Where(l => l.StartsWith(kt)).ToList()); 
+            }
         }
 
         public override void Clear()
